% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRankings.R
\name{buildRankings}
\alias{buildRankings}
\title{Generate Ranking File According to The Cell-level Scores.}
\usage{
buildRankings(Mat)
}
\arguments{
\item{Mat}{The matrix of the cell-level scores.}
}
\value{
The matrix of rankings
}
\description{
Generate Ranking File According to The Cell-level Scores.
}
\examples{
\donnttest{
buildRankings(mydata)
}
}
