% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.AUC.R
\name{calc.AUC2}
\alias{calc.AUC2}
\title{Calculate AUC for Cell Types.}
\usage{
calc.AUC2(ids, anno, cells_rankings, percent)
}
\arguments{
\item{ids}{A vector of all cell type ids to be analyzed.}

\item{anno}{The annotation table with two columns. The first column: cell_id; The second column: cell_type annotation.}

\item{cells_rankings}{The table containing the rankings of all cells, generated from buildRankings() function.}

\item{percent}{The top-rank threshold used for calculating AUC.}
}
\value{
AUC values for all cell type ids imported.
}
\description{
Calculate AUC for Cell Types.
}
\examples{
\donnttest{
calc.AUC2(id, anno, ranking, 0.05)
}
}
